refactor: move toJsonSchema logic to JsonSchema
